Add "zenpack" downloader type
Updated zenpack_versions.json to use the new "zenpack" downloader type
for all ZenPacks. Given that this is the develop branch, I went with a
configuration that will use the latest release build for all ZenPacks
except for CatalogService which will use the latest pre-release build.

See the docstring for the zenpackDownload() method in
artifact_download.py for a full description of the "zenpack" download
type's capabilities.