Properly mount source when performing a "dumpdb"
Prior to this change, performing "make dumpdb" in the devimg directory
would result in Solr failing with the following error:

  o.a.s.s.SolrDispatchFilter Could not start Solr. Check solr/home property and the logs
  o.a.s.c.SolrCore null:java.lang.IllegalArgumentException: bootstrap_confdir must be a directory of configuration files

The problem was that /opt/solr/server/solr/configsets links to
/mnt/src/modelindex/zenoss/modelindex/solr/configsets.  /mnt/src is
mounted in from the host at $SRCROOT, but $SRCROOT/modelindex does
not exist - it was moved to $SRCROOT/github.com/zenoss as part of the
zendev2 work.

Additionally remove the explicit command echo, which is unnecessary as
the script is already running with 'set -x'.